requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
av==12.3.0
backcall==0.2.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.12
decorator==5.2.0
distro==1.9.0
djitellopy==2.5.0
dynmen==0.1.5
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
fuzzywuzzy==0.18.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-rl2==1.0.5
kiwisolver==1.4.7
lap==0.5.12
Levenshtein==0.25.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
midas==0.7.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.28.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.19.0
openai==0.27.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==3.19.6
psutil==7.0.0
pure_eval==0.2.3
pvporcupine==3.0.4
pvrecorder==1.2.4
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycaw==20240210
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
pymodbus==3.6.9
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
pyxdg==0.28
PyYAML==6.0.2
pyzmq==26.2.1
rapidfuzz==3.9.7
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.20.1
rsa==4.9.1
safetensors==0.5.2
scipy==1.10.1
seaborn==0.13.2
simpleaudio==1.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
srt==3.5.3
stack-data==0.6.3
sympy==1.13.1
tello==1.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
timm==1.0.15
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==2.2.3
vosk==0.3.45
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.15.2
zipp==3.20.2
